Automatically move blockers when transform deploying.
